Moved main App component to components directory
