+++ frontend/src/components/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './css/style.css';
+import '../css/style.css';
 
 function App() {
     return (
